# remote/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Just Captions API",
    "version": "1.4.0",
    "description": "Timed captions and AI text edits. Video rendering runs locally through the open-source MCP/CLI. https://justcaptions.com/agents/"
  },
  "servers": [
    {
      "url": "https://api.justcaptions.com/v1"
    }
  ],
  "security": [
    {
      "ApiKey": []
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "get_health",
        "summary": "API liveness",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        },
        "security": []
      }
    },
    "/pricing": {
      "get": {
        "operationId": "get_pricing",
        "summary": "Current prices and free monthly allowances",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        },
        "security": []
      }
    },
    "/styles": {
      "get": {
        "operationId": "get_styles",
        "summary": "15 local rendering presets, versioned overrides, aliases and platform safe areas",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        },
        "security": []
      }
    },
    "/usage": {
      "get": {
        "operationId": "get_usage",
        "summary": "Monthly usage, estimated bill and spend cap",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        },
        "description": "Dedicated keys share root account usage, free allowance, billing plan and cap. Returns account_key and active reserved_audio_seconds/reserved_text_chars."
      }
    },
    "/transcribe": {
      "post": {
        "operationId": "post_transcribe",
        "summary": "Extracted audio to timed captions; max 12 MB decoded",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "text/vtt": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          },
          "409": {
            "description": "idempotency_conflict or request_in_progress"
          },
          "422": {
            "description": "duration_mismatch: audio exceeds measured or conservatively reserved duration; request not billed"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "audio_base64": {
                    "type": "string"
                  },
                  "mime_type": {
                    "type": "string"
                  },
                  "language": {
                    "type": "string"
                  },
                  "glossary": {
                    "type": "array",
                    "maxItems": 100,
                    "items": {
                      "type": "string"
                    }
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "json",
                      "srt",
                      "vtt"
                    ],
                    "default": "json"
                  }
                },
                "required": [
                  "audio_base64"
                ]
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "language": {
                    "type": "string"
                  },
                  "glossary": {
                    "type": "string",
                    "description": "Comma-separated glossary terms"
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "json",
                      "srt",
                      "vtt"
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,100}$"
            },
            "description": "Reuse for identical retries within 24 hours. Different input with the same key returns 409."
          }
        ]
      }
    },
    "/correct": {
      "post": {
        "operationId": "post_correct",
        "summary": "Correct up to 400 captions / 80,000 input characters; keep count and order",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          },
          "409": {
            "description": "idempotency_conflict or request_in_progress"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "captions": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 400,
                    "items": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "language": {
                    "type": "string"
                  },
                  "glossary": {
                    "type": "array",
                    "maxItems": 100,
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "captions"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,100}$"
            },
            "description": "Reuse for identical retries within 24 hours. Different input with the same key returns 409."
          }
        ]
      }
    },
    "/translate": {
      "post": {
        "operationId": "post_translate",
        "summary": "Translate captions; text edits do not realign word timing",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          },
          "409": {
            "description": "idempotency_conflict or request_in_progress"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "captions": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 400,
                    "items": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "target_language": {
                    "type": "string"
                  },
                  "glossary": {
                    "type": "array",
                    "maxItems": 100,
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "captions",
                  "target_language"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,100}$"
            },
            "description": "Reuse for identical retries within 24 hours. Different input with the same key returns 409."
          }
        ]
      }
    },
    "/emoji": {
      "post": {
        "operationId": "post_emoji",
        "summary": "One emoji per caption, null when unavailable",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          },
          "409": {
            "description": "idempotency_conflict or request_in_progress"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "captions": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 1000,
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 200
                    }
                  },
                  "language": {
                    "type": "string"
                  }
                },
                "required": [
                  "captions"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,100}$"
            },
            "description": "Reuse for identical retries within 24 hours. Different input with the same key returns 409."
          }
        ]
      }
    },
    "/signup": {
      "post": {
        "operationId": "post_signup",
        "summary": "Create a free API key, shown once; no card required",
        "responses": {
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          },
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        }
      }
    },
    "/keys/rotate": {
      "post": {
        "operationId": "post_keys_rotate",
        "summary": "Replace the API secret; the old key stops working immediately",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        }
      }
    },
    "/billing/checkout": {
      "post": {
        "operationId": "post_billing_checkout",
        "summary": "Create a Stripe checkout link; user completes payment details",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        }
      }
    },
    "/billing/portal": {
      "post": {
        "operationId": "post_billing_portal",
        "summary": "Create a Stripe billing portal link",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        }
      }
    },
    "/estimate": {
      "post": {
        "operationId": "estimate_cost",
        "summary": "Estimate incremental cost and account allowance, including concurrent holds; does not reserve credit",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "audio_seconds": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 864000
                  },
                  "text_chars": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100000000
                  }
                }
              }
            }
          }
        }
      }
    },
    "/limits": {
      "post": {
        "operationId": "set_monthly_limit",
        "summary": "Account key only: update monthly USD spend cap for new requests",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "spend_cap_cents"
                ],
                "properties": {
                  "spend_cap_cents": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10000000
                  }
                }
              }
            }
          }
        }
      }
    },
    "/keys": {
      "get": {
        "operationId": "list_agent_keys",
        "summary": "Account key only: list dedicated keys; secrets are never returned",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        }
      },
      "post": {
        "operationId": "create_agent_key",
        "summary": "Account key only: create a scoped key sharing account allowance; secret returned once",
        "responses": {
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          },
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "scopes"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "scopes": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "string",
                      "enum": [
                        "transcribe",
                        "correct",
                        "translate",
                        "emoji",
                        "usage",
                        "estimate"
                      ]
                    }
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true,
                    "description": "Defaults to 30 days; at most 366 days in future. Null disables expiry."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/keys/{key_id}/revoke": {
      "post": {
        "operationId": "revoke_agent_key",
        "summary": "Account key only: revoke a dedicated key",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request"
          },
          "401": {
            "description": "invalid_api_key"
          },
          "402": {
            "description": "payment_required"
          },
          "429": {
            "description": "quota_exceeded, free_tier_busy, spend_cap_reached or rate_limited"
          },
          "503": {
            "description": "Provider or service unavailable"
          }
        },
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^key_[A-Za-z0-9]{12}$"
            }
          }
        ]
      }
    },
    "/fonts": {
      "get": {
        "operationId": "get_fonts",
        "summary": "Named fonts, stable IDs, licenses and script coverage for local rendering",
        "security": [],
        "responses": {
          "200": {
            "description": "Font catalog. Use font_id in local MCP/CLI style overrides.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "jc_live_<32 alphanumeric characters>"
      }
    },
    "schemas": {}
  }
}